requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-tavily
PyPDF2
faiss-cpu
duckdb
pandas
python-dotenv
json5
asyncio
anthropic
google-generativeai
langgraph
langchain-core
langgraph_supervisor